+++ crates/stackable-versioned/CHANGELOG.md
@@ -9,9 +9,9 @@ All notable changes to this project will be documented in this file.
 - BREAKING: The `versioned` macro now automatically generates tests to test roundtrip converts for
   every versioned struct. This requires you to implement the `RoundtripTestData` trait for all
   structs, which returns some test data to be used for the roundtrip tests.
-  Read on the `RoundtripTestData` documentation for details ([#XXX]).
+  Read on the `RoundtripTestData` documentation for details ([#1172]).
 
-[#XXX]: https://github.com/stackabletech/operator-rs/pull/XXX
+[#1172]: https://github.com/stackabletech/operator-rs/pull/1172
 
 ## [0.8.3] - 2025-10-23